Support specififying non-default batch size
Provide new interfaces to allow calling `sendmmsg` and `recvmmsg` with a
non-default batch size, specified by API clients as a generic constant during
compile time. This is presently 1024 on Linux and FreeBSD. The default batch
size is increased to 128 on these platforms.

Also update to minimally support FreeBSD. Not all Linux features are supported
yet.